Another broken update
Modified the floorplan of the fire house
Added a door and waiting area for the ground floor and isolated the
entry so that you have to go through a door next to the secretary to
access the rest of the building
shortened the second floor so it matches better with the modified ground
floor
Modified firehouse.java, placetype.java, and worldtype.java for the new
rooms.
Made Sonja taller
Added Sonja and Steve to the game.  They should be available in the
briefing room during the day shift.