Dialogue and sex take 2
Added virginity loss descriptors for rooms in the fire house.  Thanks to
AceXP for catching that.

Changed dialogue with Sameera and Bevikar so they no longer keep saying
"name's x, nice to meet you."
Player now knows both NPCs if they are talked to
Dialogue changes after sex based on if the NPC is sexually satisfied or
not
Got rid of unneccessary stuff in BevikarDialogue

Firehouse now gets added if loading an older save.  Many thanks to AceXP
for the help with that.
Thanks again to AceXP with all the help.

things that still need to be done:
probably some sex behavior based on if dom or sub
Need to double check scheduling because something is off there
enable dialogue flags so the NPCs only do the intro once
affection stuff
Quests
finish adding NPCs
Fire events and interactions
writing