Minor structure changes
Made FIREHOUSE_FLOOR2 for the second floor of the fire house
changed NPC files so that they go to FIREHOUSE_FLOOR2 instead of
FIREHOUSE_FLOOR1 when at the second floor
put Steve on the firenightshift schedule.  He was previously on the work
schedule
Fixed firenightshift.  Thanks to AceXP for the fix
quick grammar fix in sameera dialogue xml
Added priya.xml and priyadialogue, neither are currently used.