Dialogue with AceXP contributions
Bevikar now goes to Sameera's room at SmallHours (1:00-4:00AM).  Done to
test dialogue and as a test to see how to add more of a schedule
AceXP fiddled with dialogue in firehouse.java so that NPCs can be in the
same space and you can talk to any npc with dialogue in that space
updated dialogue xml files so that NPCs have [npcname.speech] instead of
[npc.speech]
Made busy dialogue in case a NPC is 'working'
Attempted to change GenericFHResponse to GenericBusyResponse in the
office and staging area.  Still acts as GenericFHResponse
Changed Sameera's schedule so she's now at the Fire House Entrance
during an arcane storm
NPCs' schedules are now hourlyUpdate instead of turnUpdate