Sonja Boobening and other things
Changed Sonja's breast size and shape.
Added some fetishes to Sonja.  Probably going to work on those more
later.
Working with AceXP on dialogue - mostly getFHGroup in FireHouse.java to
try to detect available npcs for SMGeneric.
Fixed an oversight with the secretary after hours where you'd get stuck
in dialogue
Swapped dialogue for the secretary so the responses actually make sense
instead of you talking to an empty desk - doesn't seem to fix anything
though.
set the second floor of the fire house to be loiterable
copied gym icon to firehouse assets because I'm planning to have a gym