Tiny Bruno update
Gave Bruno body hair, set to blonde
Changed Bruno's skin to rosy. I might change it again later, not sure.
Fixed Bruno's schedule so that he goes to rooms on the second floor instead of nonexistant rooms on the ground floor
Minor changes to dialogue because I was getting an if statement error